Modifying to be a magic link only example
