+++ common.sh
@@ -73,6 +73,7 @@ options=(
     "Detection and Response Service"
     "Email OTP Authentication"
     "Identity Verification (Hosted)"
+    "Magic Link Authentication"
     "Password Authentication"
     "SMS OTP Authentication"
     "WebAuthn Passkey Authentication (autocomplete)"
@@ -83,6 +84,7 @@ optdir=(
     "password-authentication-drs"
     "login-with-email"
     "hosted-idv"
+    "login-with-magiclink"
     "password-authentication"
     "login-with-sms"
     "passkey-authentication"
