+++ login-with-magiclink/package.json
