+++ login-with-magiclink/pages/init.js
