+++ login-with-magiclink/vite.config.js
