--- login-with-mfa/pages/email-otp.html
@@ -1,150 +1,0 @@
-<!DOCTYPE html>
-<html style="display: none">
-  <head>
-    <link rel="icon" type="image/svg+xml" href="/acme-icon.svg" />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&display=swap"
-      rel="stylesheet"
-    />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>ACME</title>
-    <script type="module" src="./init.js"></script>
-  </head>
-
-  <body onload="loadPage()">
-    <div class="page">
-      <header>
-        <img src="/acme-logo.svg" />
-      </header>
-      <main>
-        <div class="content column gap">
-          <div id="emailDiv" class="column gap">
-            <h3>Enter your email</h3>
-            <input id="emailInput" type="text" size="40" placeholder="you@example.com" />
-            <button id="sendEmailButton" onclick="sendEmailOtp()" class="full-width">
-              Send Email OTP
-            </button>
-            <span style="text-align: center">or</span>
-            <button id="sendMagicLinkButton" onclick="sendMagicLink()" class="full-width">
-              Send Magic Link
-            </button>
-          </div>
-          <div id="otpDiv" class="column gap hidden">
-            <h3>Enter the OTP code</h3>
-            <input id="otpInput" type="text" size="20" />
-            <button id="loginButton" onclick="validateOtp()" class="full-width">Login</button>
-            <div id="otpText" class="text"></div>
-          </div>
-          <div id="magicLinkDiv" class="column gap hidden">
-            <div id="magicLinkText" class="text"></div>
-          </div>
-        </div>
-      </main>
-    </div>
-    <script>
-      const emailInput = document.getElementById('emailInput');
-      const otpInput = document.getElementById('otpInput');
-      const otpText = document.getElementById('otpText');
-      const otpDiv = document.getElementById('otpDiv');
-      const emailDiv = document.getElementById('emailDiv');
-      const magicLinkDiv = document.getElementById('magicLinkDiv');
-
-      function loadPage() {
-        emailDiv.classList.remove('hidden');
-        otpDiv.classList.add('hidden');
-        magicLinkDiv.classList.add('hidden');
-      }
-
-      async function sendEmailOtp() {
-        window.pageUtils.showLoading();
-        const response = await fetch('/send-email-otp', {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            email: emailInput.value,
-          }),
-        });
-
-        const status = response.status;
-        const data = await response.json();
-        console.log('Response from /send-email-otp', { status, data });
-        handleEmailOtpResponse({ status, data });
-        window.pageUtils.hideLoading();
-      }
-
-      async function sendMagicLink() {
-        window.pageUtils.showLoading();
-        const response = await fetch('/send-magic-link', {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            email: emailInput.value,
-          }),
-        });
-
-        const status = response.status;
-        const data = await response.json();
-        console.log('Response from /send-email-otp', { status, data });
-        handleMagicLinkResponse({ status, data });
-
-        window.pageUtils.hideLoading();
-      }
-
-      async function validateOtp() {
-        window.pageUtils.showLoading();
-        const response = await fetch('/verify-email-otp', {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            email: emailInput.value,
-            otpCode: otpInput.value,
-          }),
-        });
-
-        const status = response.status;
-        const data = await response.json();
-        console.log('Response from /verify-email-otp', { status, data });
-        handleVerifyOtpResponse({ status, data });
-        window.pageUtils.hideLoading();
-      }
-
-      const handleEmailOtpResponse = emailOtpResponse => {
-        console.log('Email Otp Response', emailOtpResponse);
-
-        if (emailOtpResponse?.status === 200) {
-          emailDiv.classList.add('hidden');
-          otpDiv.classList.remove('hidden');
-
-          otpText.textContent = `We sent a code to ${emailInput.value}`;
-        }
-      };
-
-      const handleMagicLinkResponse = magicLinkResponse => {
-        console.log('Magic Link Response', magicLinkResponse);
-
-        if (magicLinkResponse?.status === 200) {
-          emailDiv.classList.add('hidden');
-          magicLinkDiv.classList.remove('hidden');
-          magicLinkText.textContent = `We sent a Magic Link to ${emailInput.value}`;
-        }
-      };
-
-      const handleVerifyOtpResponse = verifyOtpResponse => {
-        console.log('Verify Otp Response', verifyOtpResponse);
-
-        if (verifyOtpResponse?.status === 200) {
-          // use the result to redirect to completion page, let it sort out if completed or not
-          window.location = verifyOtpResponse.data.result;
-        }
-      };
-    </script>
-  </body>
-</html>
