--- login-with-mfa/pages/sms-otp.html
@@ -1,112 +1,0 @@
-<!DOCTYPE html>
-<html style="display: none">
-  <head>
-    <link rel="icon" type="image/svg+xml" href="/acme-icon.svg" />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&display=swap"
-      rel="stylesheet"
-    />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>ACME</title>
-    <script type="module" src="./init.js"></script>
-  </head>
-
-  <body onload="loadPage()">
-    <div class="page">
-      <header>
-        <img src="/acme-logo.svg" />
-      </header>
-      <main>
-        <div class="content column gap">
-          <div id="phoneDiv" class="column gap">
-            <h3>MFA is Required</h3>
-            <h3>Enter your phone number</h3>
-            <input id="phoneInput" type="text" size="40" placeholder="+17346651212" />
-            <button id="sendSMSButton" onclick="sendSmsOtp()" class="full-width">
-              Send SMS OTP
-            </button>
-          </div>
-          <div id="otpDiv" class="column gap hidden">
-            <h3>Enter the OTP code</h3>
-            <input id="otpInput" type="text" size="20" />
-            <button id="loginButton" onclick="validateOtp()" class="full-width">Login</button>
-            <div id="otpText" class="text"></div>
-          </div>
-        </div>
-      </main>
-    </div>
-    <script>
-      const phoneInput = document.getElementById('phoneInput');
-      const otpInput = document.getElementById('otpInput');
-      const otpText = document.getElementById('otpText');
-      const otpDiv = document.getElementById('otpDiv');
-      const phoneDiv = document.getElementById('phoneDiv');
-
-      function loadPage() {
-        phoneDiv.classList.remove('hidden');
-        otpDiv.classList.add('hidden');
-      }
-
-      async function sendSmsOtp() {
-        window.pageUtils.showLoading();
-        const response = await fetch('/send-sms-otp', {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            phoneNumber: phoneInput.value,
-          }),
-        });
-
-        const status = response.status;
-        const data = await response.json();
-        console.log('Response from /send-sms-otp', { status, data });
-        handleSmsOtpResponse({ status, data });
-        window.pageUtils.hideLoading();
-      }
-
-      async function validateOtp() {
-        window.pageUtils.showLoading();
-        const response = await fetch('/verify-sms-otp', {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            phoneNumber: phoneInput.value,
-            otpCode: otpInput.value,
-          }),
-        });
-
-        const status = response.status;
-        const data = await response.json();
-        console.log('Response from /verify-sms-otp', { status, data });
-        handleVerifyOtpResponse({ status, data });
-        window.pageUtils.hideLoading();
-      }
-
-      const handleSmsOtpResponse = smsOtpResponse => {
-        console.log('SMS Otp Response', smsOtpResponse);
-
-        if (smsOtpResponse?.status === 200) {
-          phoneDiv.classList.add('hidden');
-          otpDiv.classList.remove('hidden');
-
-          otpText.textContent = `We sent a code to ${phoneInput.value}`;
-        }
-      };
-
-      const handleVerifyOtpResponse = verifyOtpResponse => {
-        console.log('Verify Otp Response', verifyOtpResponse);
-
-        if (verifyOtpResponse?.status === 200) {
-          // use the result to redirect to completion page, let it sort out if completed or not
-          window.location = verifyOtpResponse.data.result;
-        }
-      };
-    </script>
-  </body>
-</html>
